+++ Source/merchant/systems/Dnd5eCurrencyCalculator.ts
@@ -8,15 +8,15 @@ import {ItemData} from "@league-of-foundry-developers/foundry-vtt-types/src/foun
 import * as Console from "console";
 import Globals from "../../Globals";
 
-let conversionRates = {"pp": 1,
-	"gp": 1,
-	"ep": 1,
-	"sp": 1,
-	"cp": 1
+let conversionRates: {[key:string]: number} = {"pp": 1,
+	"gp": 10,
+	"ep": 2,
+	"sp": 5,
+	"cp": 10
 };
 
 
-const compensationCurrency = {"pp": "gp", "gp": "ep", "ep": "sp", "sp": "cp"};
+const convertFromHigherCurrency: {[key:string]: string} = {"cp": "sp", "sp": "ep", "ep": "gp", "gp": "pp"};
 
 
 export default class Dnd5eCurrencyCalculator extends CurrencyCalculator {
@@ -78,10 +78,10 @@ export default class Dnd5eCurrencyCalculator extends CurrencyCalculator {
 
     buyerHaveNotEnoughFunds(itemCostInGold:number, buyerFunds: any) {
 
-        let itemCostInPlatinum = itemCostInGold / conversionRates["gp"]
-        let buyerFundsAsPlatinum = this.convertToPlatinum(buyerFunds);
+        let itemCostInCopper = this.convertCurrencyToLowest({gp: itemCostInGold})
+        let buyerFundsAsCopper = this.convertCurrencyToLowest(buyerFunds);
 
-        return itemCostInPlatinum > buyerFundsAsPlatinum;
+        return itemCostInCopper > buyerFundsAsCopper;
     }
 
     convertToPlatinum(buyerFunds: any) {
@@ -99,165 +99,126 @@ export default class Dnd5eCurrencyCalculator extends CurrencyCalculator {
         buyer.update({ "data.currency": buyerFunds });
     }
 
+	convertCurrencyToLowest(buyerFunds: any): number {
+		let buyerFundsAsCopper = 0;
+		if (buyerFunds["cp"]) {
+			buyerFundsAsCopper += buyerFunds["cp"];
+		}
+		if (buyerFunds["sp"]) {
+			buyerFundsAsCopper += buyerFunds["sp"] * conversionRates["cp"];
+		}
+		if (buyerFunds["ep"]) {
+			buyerFundsAsCopper += buyerFunds["ep"] * conversionRates["sp"] * conversionRates["cp"];
+		}
+		if (buyerFunds["gp"]) {
+			buyerFundsAsCopper += buyerFunds["gp"] * conversionRates["ep"] * conversionRates["sp"] * conversionRates["cp"];
+		}
+		if (buyerFunds["pp"]) {
+			buyerFundsAsCopper += buyerFunds["pp"] * conversionRates["gp"] * conversionRates["ep"] * conversionRates["sp"] * conversionRates["cp"];
+		}
+		return buyerFundsAsCopper
+	}
 
     subtractAmountFromActor(buyer: Actor, buyerFunds: any, itemCostInGold: number) {
-
-        let itemCostInPlatinum = itemCostInGold / conversionRates["gp"]
-        let buyerFundsAsPlatinum = this.convertToPlatinum(buyerFunds);
-
-        console.log(`buyerFundsAsPlatinum : ${buyerFundsAsPlatinum}`);
-
-        let convertCurrency = (<Game>game).settings.get(Globals.ModuleName, "convertCurrency");
-
-        if (convertCurrency) {
-            buyerFundsAsPlatinum -= itemCostInPlatinum;
-
-            // Remove every coin we have
-            for (let currency in buyerFunds) {
-                buyerFunds[currency] = 0
-            }
-
-            // Give us fractions of platinum coins, which will be smoothed out below
-            buyerFunds["pp"] = buyerFundsAsPlatinum
-
-        } else {
-            // We just pay in partial platinum.
-            // We dont care if we get partial coins or negative once because we compensate later
-            buyerFunds["pp"] -= itemCostInPlatinum
-
-            // Now we exchange all negative funds with coins of lower value
-            // We dont need to care about running out of money because we checked that earlier
-            for (let currency in buyerFunds) {
-                let amount = buyerFunds[currency]
-                // console.log(`${currency} : ${amount}`);
-                if (amount >= 0) continue;
-
-                // If we have ever so slightly negative cp, it is likely due to floating point error
-                // We dont care and just give it to the player
-                if (currency == "cp") {
-                    buyerFunds["cp"] = 0;
-                    continue;
-                }
-
-                // @ts-ignore
-				let compCurrency = compensationCurrency[currency]
-
-                buyerFunds[currency] = 0;
-                // @ts-ignore
-				buyerFunds[compCurrency] += amount * conversionRates[compCurrency]; // amount is a negative value so we add it
-                // console.log(`Substracted: ${amount * conversionRates[compCurrency]} ${compCurrency}`);
-            }
-        }
-
-        // console.log(`Smoothing out`);
-        // Finally we exchange partial coins with as little change as possible
-        for (let currency in buyerFunds) {
-            let amount = buyerFunds[currency]
-
-            // console.log(`${currency} : ${amount}: ${conversionRates[currency]}`);
-
-            // We round to 5 decimals. 1 pp is 1000cp, so 5 decimals always rounds good enough
-            // We need to round because otherwise we get 15.99999999999918 instead of 16 due to floating point precision
-            // If we would floor 15.99999999999918 everything explodes
-            let newFund = Math.floor(Math.round(amount * 1e5) / 1e5);
-            buyerFunds[currency] = newFund;
-
-            // console.log(`New Buyer funds ${currency}: ${buyerFunds[currency]}`);
-            // @ts-ignore
-			let compCurrency = compensationCurrency[currency]
-
-            // We dont care about fractions of CP
-            if (currency != "cp") {
-                // We calculate the amount of lower currency we get for the fraction of higher currency we have
-                // @ts-ignore
-				let toAdd = Math.round((amount - newFund) * 1e5) / 1e5 * conversionRates[compCurrency]
-                buyerFunds[compCurrency] += toAdd
-                // console.log(`Added ${toAdd} to ${compCurrency} it is now ${buyerFunds[compCurrency]}`);
-            }
-        }
-        // buyerFunds = buyerFunds - itemCostInGold;
+		let useEP = (<Game>game).settings.get(Globals.ModuleName, "useEP");
+		buyerFunds = this.calculateNewBuyerFunds(itemCostInGold, buyerFunds);
+		if (!useEP) {
+			this.convertEP(buyerFunds);
+		}
+		// buyerFunds = buyerFunds - itemCostInGold;
         this.updateActorWithNewFunds(buyer,buyerFunds);
         console.log(`Merchant sheet | Funds after purchase: ${buyerFunds}`);
     }
 
-    addAmountForActor(seller: Actor, sellerFunds: any, itemCostInGold: number) {
+	public convertEP(buyerFunds: any): any {
+		if (buyerFunds["ep"] > 0) {
+			let restEP = buyerFunds["ep"] % conversionRates["ep"];
+			if (restEP > 0) {
+				buyerFunds["sp"] = restEP * conversionRates["sp"];
+				buyerFunds["ep"] -= restEP;
+			}
+			let higherFund = Math.floor(buyerFunds["ep"] / conversionRates["ep"]);
+			if (higherFund > 0) {
+				buyerFunds["gp"] += higherFund
+				buyerFunds["ep"] = 0;
+			}
+		}
+		return buyerFunds;
+	}
 
-        let itemCostInPlatinum = itemCostInGold / conversionRates["gp"]
-        let buyerFundsAsPlatinum = this.convertToPlatinum(sellerFunds);
+	public calculateNewBuyerFunds(itemCostInGold: number, funds: any): any {
+		let itemCostInCopper = this.convertCurrencyToLowest({gp: itemCostInGold})
+		let buyerFundsInCopper = this.convertCurrencyToLowest(funds);
+		if ((buyerFundsInCopper - itemCostInCopper) < 0) {
+			throw "Could not do the transaction"
+		}
+		console.log(`buyerFundsInCopper : ${buyerFundsInCopper}`);
 
-        console.log(`buyerFundsAsPlatinum : ${buyerFundsAsPlatinum}`);
+		let buyerFunds = funds;
+		buyerFunds["cp"] -= itemCostInCopper
+		this.subtractAndConvertToHigherFund(buyerFunds, "cp");
+		return buyerFunds;
+	}
 
-        let convertCurrency = (<Game>game).settings.get(Globals.ModuleName, "convertCurrency");
+	public calculatePriceToBuyerFunds(itemCostInGold: number): any {
+		let itemCostInCopper = this.convertCurrencyToLowest({gp: itemCostInGold})
 
-        if (convertCurrency) {
-            buyerFundsAsPlatinum += itemCostInPlatinum;
+		let buyerFunds = {pp: 0, gp: 0, ep: 0, sp: 0, cp: itemCostInCopper};
+		this.convertToHigherFund(buyerFunds, "cp");
+		return buyerFunds;
+	}
 
-            // Remove every coin we have
-            for (let currency in sellerFunds) {
-                sellerFunds[currency] = 0
-            }
+	private convertToHigherFund(buyerFunds:any, currency: string) {
+		let higherCurrency: string = convertFromHigherCurrency[currency];
+		if (higherCurrency) {
+			let higherCurrencyNeeded = Math.floor((buyerFunds[currency]) / conversionRates[currency])
+			let rest = (buyerFunds[currency]) % conversionRates[currency];
+			if (rest != 0) {
+				buyerFunds[currency] = rest;
+			} else {
+				buyerFunds[currency] = 0
+			}
+			if (higherCurrencyNeeded > 0) {
+				buyerFunds[higherCurrency] = higherCurrencyNeeded;
+				this.convertToHigherFund(buyerFunds,higherCurrency);
+			}
+		}
+		return buyerFunds;
+	}
 
-            // Give us fractions of platinum coins, which will be smoothed out below
-            sellerFunds["pp"] = buyerFundsAsPlatinum
+	private subtractAndConvertToHigherFund(buyerFunds: any, currency: string) {
+		let higherCurrency: string = convertFromHigherCurrency[currency];
+		if (higherCurrency && buyerFunds[currency] < 0) {
+			let higherCurrencyNeeded = Math.floor((-1 * buyerFunds[currency]) / conversionRates[currency])
+			let rest = (-1 * buyerFunds[currency]) % conversionRates[currency];
+			if (rest != 0) {
+				higherCurrencyNeeded += 1
+				buyerFunds[currency] = (conversionRates[currency] - rest);
+			} else {
+				buyerFunds[currency] = 0
+			}
+			buyerFunds[higherCurrency] -= higherCurrencyNeeded;
 
-        } else {
-            // We just pay in partial platinum.
-            // We dont care if we get partial coins or negative once because we compensate later
-            sellerFunds["pp"] += itemCostInPlatinum
+			if (buyerFunds[higherCurrency] < 0) {
+				this.subtractAndConvertToHigherFund(buyerFunds, higherCurrency);
+			}
+		}
+	}
 
-            // Now we exchange all negative funds with coins of lower value
-            // We dont need to care about running out of money because we checked that earlier
-            for (let currency in sellerFunds) {
-                let amount = sellerFunds[currency]
-                // console.log(`${currency} : ${amount}`);
-                if (amount >= 0) continue;
+	addAmountForActor(seller: Actor, sellerFunds: any, itemCostInGold: number) {
+		let itemCostInCopper = this.convertCurrencyToLowest({gp: itemCostInGold})
 
-                // If we have ever so slightly negative cp, it is likely due to floating point error
-                // We dont care and just give it to the player
-                if (currency == "cp") {
-                    sellerFunds["cp"] = 0;
-                    continue;
-                }
-
-                // @ts-ignore
-				let compCurrency = compensationCurrency[currency]
-
-                sellerFunds[currency] = 0;
-                // @ts-ignore
-				sellerFunds[compCurrency] += amount * conversionRates[compCurrency]; // amount is a negative value so we add it
-                // console.log(`Substracted: ${amount * conversionRates[compCurrency]} ${compCurrency}`);
-            }
-        }
-
-        // console.log(`Smoothing out`);
-        // Finally we exchange partial coins with as little change as possible
-        for (let currency in sellerFunds) {
-            let amount = sellerFunds[currency]
-
-            // console.log(`${currency} : ${amount}: ${conversionRates[currency]}`);
-
-            // We round to 5 decimals. 1 pp is 1000cp, so 5 decimals always rounds good enough
-            // We need to round because otherwise we get 15.99999999999918 instead of 16 due to floating point precision
-            // If we would floor 15.99999999999918 everything explodes
-            let newFund = Math.floor(Math.round(amount * 1e5) / 1e5);
-            sellerFunds[currency] = newFund;
-
-            // console.log(`New Buyer funds ${currency}: ${sellerFunds[currency]}`);
-            // @ts-ignore
-			let compCurrency = compensationCurrency[currency]
-
-            // We dont care about fractions of CP
-            if (currency != "cp") {
-                // We calculate the amount of lower currency we get for the fraction of higher currency we have
-                // @ts-ignore
-				let toAdd = Math.round((amount - newFund) * 1e5) / 1e5 * conversionRates[compCurrency]
-                sellerFunds[compCurrency] += toAdd
-                // console.log(`Added ${toAdd} to ${compCurrency} it is now ${sellerFunds[compCurrency]}`);
-            }
-        }
-        // sellerFunds = sellerFunds - itemCostInGold;
-        this.updateActorWithNewFunds(seller,sellerFunds);
-        console.log(`Merchant Sheet | Funds after sell: ${sellerFunds}`);
+		let convertToHigherFund = this.convertToHigherFund({pp:0,gp:0,ep:0,sp:0,cp:itemCostInCopper},"cp");
+		sellerFunds["cp"] += convertToHigherFund["cp"]
+		sellerFunds["sp"] += convertToHigherFund["sp"]
+		sellerFunds["ep"] += convertToHigherFund["ep"]
+		sellerFunds["gp"] += convertToHigherFund["gp"]
+		sellerFunds["pp"] += convertToHigherFund["pp"]
+		let useEP = (<Game>game).settings.get(Globals.ModuleName, "useEP");
+		if (!useEP) {
+			this.convertEP(sellerFunds);
+		}
+		this.updateActorWithNewFunds(seller,sellerFunds);
     }
 
     priceInText(itemCostInGold: number): string {
@@ -332,14 +293,6 @@ export default class Dnd5eCurrencyCalculator extends CurrencyCalculator {
 	}
 
 	public initSettings() {
-		(<Game>game).settings.register(Globals.ModuleName, "convertCurrency", {
-            name: "Convert currency after purchases?",
-            hint: "If enabled, all currency will be converted to the highest denomination possible after a purchase. If disabled, currency will subtracted simply.",
-            scope: "world",
-            config: true,
-            default: false,
-            type: Boolean
-        });
 
 		conversionRates = {"pp": 1,
 			// @ts-ignore
@@ -354,7 +307,18 @@ export default class Dnd5eCurrencyCalculator extends CurrencyCalculator {
 		super.initSettings();
     }
 
-    getPriceFromItem(item: Item) {
+	public registerSystemSettings() {
+		(<Game>game).settings.register(Globals.ModuleName, "useEP", {
+			name: "Use EP in currency conversion?",
+			hint: "If disabled, EP will not be used and will remove any EP in the purchase or selling .",
+			scope: "world",
+			config: true,
+			default: true,
+			type: Boolean
+		});
+	}
+
+	getPriceFromItem(item: Item) {
         // @ts-ignore
 		return item.data.price;
     }
@@ -366,4 +330,6 @@ export default class Dnd5eCurrencyCalculator extends CurrencyCalculator {
 	currency(): string {
 		return 'GP';
 	}
+
+
 }